Project2/requirements.txt
appnope==0.1.0
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
cycler==0.10.0
decorator==4.1.2
enum34==1.1.6
ipdb==0.10.3
ipython==5.5.0
ipython-genutils==0.2.0
matplotlib==2.1.1
numpy==1.13.3
olefile==0.44
pathlib2==2.3.0
pexpect==4.3.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
scandir==1.6
simplegeneric==0.8.1
six==1.11.0
subprocess32==3.2.7
traitlets==4.3.2
wcwidth==0.1.7
wordcloud==1.3.1
